updates 12-4-2020
